+++ packages/theme-tools/src/index.ts
@@ -1,4 +1,2 @@
-export * from './color'
-export * from './css-vars'
 export * from './style-alias'
-export * from './types'
+export * from './color'