+++ packages/theme/src/index.ts
@@ -1,5 +1,6 @@
 export * from './border'
 export * from './colors'
+export * from './css-vars'
 export * from './shadows'
 export * from './theme'
 export * from './types'